Zen mode to follow global slice char setting
In our corp we have the most useful info at the front of the friendly name, therefore it is not handy to cut everything except the last 6 chars by default in Zen mode. I can also not think of a reason why the global setting would not suffice. Thus I´m suggesting to change it to follow the global sliceLastChars setting, just like the normal map. 

Change is (once again) not tested, but it should work.